Plant-Disease-Detection/requirements.txt
# Plant Disease Detection System - Requirements
# Updated for AI Project with Agent, Search Algorithms, and Baseline Models

# Core Dependencies
numpy>=1.20.2
pandas>=1.2.4
Pillow>=8.2.0
matplotlib>=3.4.2
seaborn>=0.11.1
scipy>=1.6.3

# Machine Learning - Scikit-learn
scikit-learn>=0.24.2
joblib>=1.0.1

# Deep Learning - PyTorch
torch>=1.8.1
torchvision>=0.9.1

# Flask Web Application
Flask>=1.1.2
gunicorn>=20.1.0
Werkzeug>=1.0.1
Jinja2>=2.11.3
MarkupSafe>=1.1.1
itsdangerous>=1.1.0
click>=7.1.2

# Utilities
python-dateutil>=2.8.1
pytz>=2021.1
six>=1.15.0
typing-extensions>=3.7.4.3

# Jupyter and Notebook Support
jupyter>=1.0.0
notebook>=6.4.0
ipykernel>=5.5.5
ipywidgets>=7.6.3

# Visualization
plotly>=5.1.0
opencv-python>=4.5.2

# Model Serialization
h5py>=3.2.1

# Progress Bars
tqdm>=4.61.1

# Configuration
pyyaml>=5.4.1

# Testing (Optional)
pytest>=6.2.4
pytest-cov>=2.12.1

# Code Quality (Optional)
black>=21.6b0
flake8>=3.9.2
